requirements.txt
streamlit
pandas
pyyaml
requests
python-dotenv